Implement order status update and retrieval
The commit adds functionality to the ShopService class to allow for updating an order's status and retrieving orders by status. In addition to this, the Lombok library has been introduced and used to enhance the Order class and simplify the code. The relevant test cases have been added to the ShopServiceTest class.