Update Main.java to create and add orders to ShopService
The code modifications in Main.java revolve around the creation of new products and the addition of these products to orders in the ShopService. The products are defined individually, making them easier to reference when creating orders. This change provides an example of how orders can be created and added to ShopService, enhancing usage understanding.