Introduce IdService for generating UUIDs
A new IdService class was created for generating UUIDs, replacing direct calls to UUID.randomUUID() in the ShopService class. This service has also been incorporated in corresponding tests to ensure the new functionality is properly utilized and tested. This change centralizes the generation of UUIDs and provides opportunities for more flexible id generation strategies in the future.